+++ akash/client/providers-api/providers.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
+	"net/url"
 	"terraform-provider-akash/akash/client/types"
 )
 
@@ -33,7 +34,8 @@ func New(host string) *ProvidersClient {
 
 // GetAllProviders gets all the providers from the providers' API. Returns error in case something goes wrong.
 func (c *ProvidersClient) GetAllProviders() ([]types.Provider, error) {
-	req, err := http.NewRequest(http.MethodGet, c.host, nil)
+	addr, _ := url.JoinPath(c.host, "provider/")
+	req, err := http.NewRequest(http.MethodGet, addr, nil)
 	if err != nil {
 		return nil, err
 	}